Fix reading packet SMSG_PLAYER_STORAGE_REMOVE (tmwa)

+++ src/net/tmwa/inventoryrecv.cpp
@@ -481,7 +481,7 @@ void InventoryRecv::processPlayerStorageRemove(Net::MessageIn &msg)
     BLOCK_START("InventoryRecv::processPlayerStorageRemove")
     // Move an item out of storage
     const int index = msg.readInt16("index") - STORAGE_OFFSET;
-    const int amount = msg.readInt16("amount");
+    const int amount = msg.readInt32("amount");
     if (Ea::InventoryRecv::mStorage)
     {
         if (Item *const item = Ea::InventoryRecv::mStorage->getItem(index))
